# services/identity/openapi.yaml
openapi: 3.0.3
info:
  title: QARoom identity-service
  version: 0.0.0
  description: Users, communities-as-tenants, memberships, sessions, and JWT/JWKS. Generated from Zod — do not edit by hand.
paths:
...
  /ws/tickets:
    post:
      operationId: createWsTicket
      summary: Mint a one-use WebSocket handshake ticket
      description: Issues a one-use, 30-second ticket bound to the authenticated principal. Present it in the WebSocket subprotocol (`ticket.<ticket>`). Deliberately not idempotent — each call mints a fresh ticket (ADR-0013).
      responses:
        "201":
          description: A freshly minted ticket.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TicketResponse"
              example:
                ticket: tkt_01HZY0K7M3QF8VN2J5RX9TB4CK
                expires_in_seconds: 30
        "401":
          description: A valid Bearer access token is required.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: https://qaroom.dev/errors/missing-bearer-token
                title: Authentication failed
                status: 401
                detail: Authentication failed.
                instance: /ws/tickets
                retryable: false
                next_actions: []
                failure_domain: authentication
      tags:
        - ws
      parameters:
        - name: Authorization
          in: header
          required: true
          description: Bearer access token (the JWT minted by createSession).
          schema:
            type: string
            pattern: ^Bearer .+$
  /ws/tickets/redeem:
    post:
      operationId: redeemWsTicket
      summary: Redeem a WebSocket ticket (internal, gateway → identity)
      description: Consumes a ticket exactly once and returns the principal it authorizes. Called by the gateway before upgrading a WebSocket connection. An unknown, expired, or already-redeemed ticket is a 401.
      responses:
        "200":
          description: The principal the ticket authorizes.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RedeemTicketResponse"
              example:
                user_id: user_01HZY0K7M3QF8VN2J5RX9TB4CF
                memberships:
                  - community_id: comm_01HZY0K7M3QF8VN2J5RX9TB4CD
                    role: member
        "401":
          description: The ticket is unknown, expired, or already redeemed.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              example:
                type: https://qaroom.dev/errors/ticket-invalid
                title: Authentication failed
                status: 401
                detail: Authentication failed.
                instance: /ws/tickets/redeem
                retryable: false
                next_actions: []
                failure_domain: authentication
      tags:
        - ws
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemTicketRequest"
            example:
              ticket: tkt_01HZY0K7M3QF8VN2J5RX9TB4CK
...
components:
  schemas:
...
    CommunityId:
      type: string
      pattern: ^comm_[0-9A-HJKMNP-TV-Z]{26}$
      description: "Branded identifier: comm_<ULID>"
...
    FailureDomain:
      type: string
      enum:
        - validation
        - authentication
        - authorization
        - tenant_resolution
        - rate_limit
        - conflict
        - not_found
        - dependency_failure
        - internal_error
...
    MembershipClaim:
      type: object
      properties:
        community_id:
          $ref: "#/components/schemas/CommunityId"
        role:
          $ref: "#/components/schemas/Role"
      required:
        - community_id
        - role
      additionalProperties: false
      description: A {community_id, role} pair in the access token.
    NextAction:
      type: object
      properties:
        verb:
          type: string
          enum:
            - GET
            - POST
            - PUT
            - PATCH
            - DELETE
        href:
          type: string
        description:
          type: string
      required:
        - verb
        - href
        - description
      additionalProperties: false
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          description: Absolute URI identifying the problem type.
        title:
          type: string
        status:
          type: integer
          minimum: 100
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        retryable:
          type: boolean
        next_actions:
          type: array
          items:
            $ref: "#/components/schemas/NextAction"
        failure_domain:
          $ref: "#/components/schemas/FailureDomain"
      required:
        - type
        - title
        - status
        - retryable
        - next_actions
        - failure_domain
      additionalProperties: false
      description: RFC 7807 Problem Details with QARoom agent extensions.
    RedeemTicketRequest:
      type: object
      properties:
        ticket:
          $ref: "#/components/schemas/TicketId"
      required:
        - ticket
      additionalProperties: false
      description: Body for the internal ticket redeem call.
    RedeemTicketResponse:
      type: object
      properties:
        user_id:
          $ref: "#/components/schemas/UserId"
        memberships:
          type: array
          items:
            $ref: "#/components/schemas/MembershipClaim"
      required:
        - user_id
        - memberships
      additionalProperties: false
      description: The principal a redeemed ticket authorizes.
    Role:
      type: string
      enum:
        - owner
        - moderator
        - member
      description: Membership role within a community.
...
    TicketId:
      type: string
      pattern: ^tkt_[0-9A-HJKMNP-TV-Z]{26}$
      description: "Branded identifier: tkt_<ULID>"
    TicketResponse:
      type: object
      properties:
        ticket:
          $ref: "#/components/schemas/TicketId"
        expires_in_seconds:
          type: integer
          minimum: 0
          exclusiveMinimum: true
          maximum: 9007199254740991
      required:
        - ticket
        - expires_in_seconds
      additionalProperties: false
      description: A freshly minted WebSocket handshake ticket.
...
    UserId:
      type: string
      pattern: ^user_[0-9A-HJKMNP-TV-Z]{26}$
      description: "Branded identifier: user_<ULID>"
servers:
  - url: http://localhost:8082
    description: local docker-compose